Add locks and move transactions for addBlock
The goal here is to get rid of the global addBlock transaction. The
issue is that by wrapping the entire addBlock in a transaction, it can't
deal with non transactional in memory state that's related to just
individual block operations.

As a refreshser, in a single addBlock() call many block operations can
be called, like in a reorganize where we disconnect, and reconnect many
blocks before actually adding back the new head block.